+++ RefactoringsExperiments/src/main/java/org/jetbrains/research/extractMethodExperiments/legacy/LegacyFalseRefactoringsExtractor.java
@@ -21,6 +21,7 @@ import org.jetbrains.research.extractMethodExperiments.csv.SparseCSVBuilder;
 import org.jetbrains.research.extractMethodExperiments.csv.models.CSVItem;
 import org.jetbrains.research.extractMethodExperiments.csv.models.Feature;
 import org.jetbrains.research.extractMethodExperiments.utils.calcers.*;
+import org.jetbrains.research.extractMethodExperiments.utils.feature.generators.*;
 import org.refactoringminer.api.GitService;
 import org.refactoringminer.util.GitServiceImpl;
