+++ RefactoringsExperiments/src/main/java/org/jetbrains/research/extractMethodExperiments/utils/feature/generators/GitBlameAnalyzer.java
@@ -1,4 +1,4 @@
-package org.jetbrains.research.extractMethodExperiments.utils.calcers;
+package org.jetbrains.research.extractMethodExperiments.utils.feature.generators;
 
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.errors.GitAPIException;
