+++ RefactoringsExperiments/src/main/java/org/jetbrains/research/extractMethodExperiments/utils/feature/generators/KeywordsCalculator.java
@@ -1,4 +1,4 @@
-package org.jetbrains.research.extractMethodExperiments.utils.calcers;
+package org.jetbrains.research.extractMethodExperiments.utils.feature.generators;
 
 import org.jetbrains.research.extractMethodExperiments.csv.SparseCSVBuilder;
 import org.jetbrains.research.extractMethodExperiments.csv.models.CSVItem;
