+++ RefactoringsExperiments/src/test/java/org/jetbrains/research/extractMethodExperiments/models/FragmentTest.java
@@ -1,8 +1,6 @@
-package org.jetbrains.research.extractMethodExperiments.code_models;
+package org.jetbrains.research.extractMethodExperiments.models;
 
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class FragmentTest {
     @Test
